update import path for moved enum package
