+++ internal/test_utils/datacenter_test_objects/widgets.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/Juniper/apstra-go-sdk/apstra"
-	"github.com/Juniper/apstra-go-sdk/apstra/enum"
+	"github.com/Juniper/apstra-go-sdk/enum"
 	timeutils "github.com/Juniper/apstra-go-sdk/internal/time_utils"
 	"github.com/stretchr/testify/require"
 )
